backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = ""
authors = [
    {name = "SimHoZebs",email = "simhozebs@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django (>=5.2.6,<6.0.0)",
    "psycopg2-binary (>=2.9.0,<3.0.0)",
    "djangorestframework (>=3.15.0,<4.0.0)",
    "google-cloud-speech (>=2.33.0,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
